+++ src/mockService.js
@@ -1,5 +1,3 @@
-'use strict';
-
 Pact.MockService = Pact.MockService || {};
 
 (function() {
@@ -9,6 +7,12 @@ Pact.MockService = Pact.MockService || {};
     var _baseURL = 'http://127.0.0.1:' + opts.port;
     var _interactions = [];
 
+    if (typeof(opts.done) !== 'function') {
+      throw new Error('Error creating MockService. Please provide an option called "done", that is a function that asserts (using your test framework of choice) that the first argument, error, is null.');
+    }
+
+    var _doneCallback = opts.done;
+
     var _pactDetails = {
       consumer: {
         name: opts.consumer
@@ -16,6 +20,50 @@ Pact.MockService = Pact.MockService || {};
       provider: {
         name: opts.provider
       }
+    };
+
+    var setupInteractionsSequentially = function(interactions, index, callback) {
+      if (index >= interactions.length) {
+        callback();
+        return;
+      }
+
+      Pact.MockServiceRequests.postInteraction(interactions[index], _baseURL, function(error) {
+        if (error) {
+          callback(error);
+          return;
+        }
+
+        setupInteractionsSequentially(interactions, index + 1, callback);
+      });
+    };
+
+    var cleanAndSetup = function(callback) {
+      // Cleanup the interactions from the previous test
+      Pact.MockServiceRequests.deleteInteractions(_baseURL, function(deleteInteractionsError) {
+        if (deleteInteractionsError) {
+          callback(deleteInteractionsError);
+          return;
+        }
+
+        // Post the new interactions
+        var interactions = _interactions;
+        _interactions = []; //Clean the local setup
+        setupInteractionsSequentially(interactions, 0, callback);
+      });
+    };
+
+    var verifyAndWrite = function(callback) {
+      //Verify that the expected interactions have occurred
+      Pact.MockServiceRequests.getVerification(_baseURL, function(verifyError) {
+        if (verifyError) {
+          callback(verifyError);
+          return;
+        }
+
+        //Write the pact file
+        Pact.MockServiceRequests.postPact(_pactDetails, _baseURL, callback);
+      });
     };
 
     this.given = function(providerState) {
@@ -30,69 +78,25 @@ Pact.MockService = Pact.MockService || {};
       return interaction;
     };
 
-    this.clean = function() {
-      var response = this.request('DELETE', _baseURL + '/interactions', null);
-      if (_isNotSuccess(response)) {
-        throw 'pact-consumer-js-dsl: Pact cleanup failed. ' + response.responseText;
-      }
-    };
-
-    this.setup = function() {
-      var interactions = _interactions;
-      _interactions = []; //Clean the local setup
-      var self = this;
-
-      interactions.forEach(function(interaction) {
-        var response = self.request('POST', _baseURL + '/interactions', JSON.stringify(interaction));
-        if (_isNotSuccess(response)) {
-          throw 'pact-consumer-js-dsl: Pact interaction setup failed. ' + response.responseText;
-        }
-      });
-    };
-
-    this.verify = function() {
-      var response = this.request('GET', _baseURL + '/interactions/verification', null);
-      if (_isNotSuccess(response)) {
-        throw 'pact-consumer-js-dsl: Pact verification failed. ' + response.responseText;
-      }
-    };
-
-    this.write = function() {
-      var response = this.request('POST', _baseURL + '/pact', JSON.stringify(_pactDetails));
-      if (_isNotSuccess(response)) {
-        throw 'pact-consumer-js-dsl: Could not write the pact file. ' + response.responseText;
-      }
-    };
-
-    this.run = function(testFunction) {
-      var self = this;
-      self.clean(); // Cleanup the interactions from the previous test
-      self.setup(); // Post the new interactions
-
-      var runComplete = function(testComplete) {
-        self.verify(); //Verify that the expected interactions have occurred
-        self.write();  //Write the pact file
-
-        if (typeof(testComplete) === 'function') {
-          testComplete();
-        }
+    this.run = function(onRunComplete, testFunction) {
+      var done = function (error) {
+        _doneCallback(error);
+        onRunComplete();
       };
 
-      testFunction(runComplete); // Call the tests
-    };
+      cleanAndSetup(function(error) {
+        if (error) {
+          done(error);
+          return;
+        }
 
-    this.request = function(method, url, body) {
-      var response = new XMLHttpRequest();
-      response.open(method, url, false);
-      response.setRequestHeader('X-Pact-Mock-Service', 'true');
-      response.setRequestHeader('Content-type', 'application/json');
-      response.send(body);
-      return response;
-    };
+        var runComplete = function() {
+          verifyAndWrite(done);
+        };
 
-    function _isNotSuccess(response) {
-      return parseInt(response.status, 10) !== 200;
-    }
+        testFunction(runComplete); // Call the tests
+      });
+    };
   }
 
   this.create = function(opts) {